+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Gluecodium project Release Notes
 
+## Unreleased
+### Bug fixes:
+  * `@Deprecated` attribute specified on a property in IDL is now propagated to that property's accessors (but only if
+    they lack their own attribute of this type).
+
 ## 8.6.8
 Release date: 2020-12-16
 ### Bug fixes: